# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da API do Press Ticket®",
    "description": "Documentação da API para envio de mensagens e gerenciamento de contatos.",
    "contact": {
      "email": "robson.tenorio@gmail.com"
    },
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://devapi.pressticket.com.br",
      "description": "API Press Ticket®"
    }
  ],
  "paths": {
...
    "/v1/tickets": {
      "get": {
        "summary": "Listar Tickets",
        "description": "Listar todos os tickets via API\n\nPermissões requeridas: `read:tickets`",
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "apiToken": [
              "read:tickets"
            ]
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Status do ticket (open, pending, closed)",
            "schema": {
              "type": "string",
              "enum": ["open", "pending", "closed"]
            }
          },
          {
            "name": "searchParam",
            "in": "query",
            "description": "Termo para busca",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Número da página para paginação",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Data inicial para filtro (formato YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Data final para filtro (formato YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "showAll",
            "in": "query",
            "description": "Mostrar todos os tickets (true/false)",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "queueIds",
            "in": "query",
            "description": "IDs dos setores para filtrar, formato JSON array ou separados por vírgula",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "withUnreadMessages",
            "in": "query",
            "description": "Filtrar tickets com mensagens não lidas (true/false)",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "all",
            "in": "query",
            "description": "Retornar todos os tickets sem paginação (true/false)",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tickets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tickets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Ticket"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Total de tickets encontrados"
                    },
                    "hasMore": {
                      "type": "boolean",
                      "description": "Indica se há mais páginas de resultados"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação dos parâmetros"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'read:tickets'"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      }
    },
    "/v1/tickets/{ticketId}": {
      "get": {
        "summary": "Obter Ticket",
        "description": "Obter detalhes de um ticket específico via API\n\nPermissões requeridas: `read:tickets`",
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "apiToken": [
              "read:tickets"
            ]
          }
        ],
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "ID do ticket",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket encontrado"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'read:tickets'"
          },
          "404": {
            "description": "Ticket não encontrado"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      },
      "put": {
        "summary": "Atualizar Ticket",
        "description": "Atualizar um ticket específico via API\n\nPermissões requeridas: `update:tickets`",
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "apiToken": [
              "update:tickets"
            ]
          }
        ],
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "ID do ticket",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "Status do ticket (open, pending, closed)",
                    "enum": ["open", "pending", "closed"]
                  },
                  "queueId": {
                    "type": "integer",
                    "description": "ID do setor"
                  },
                  "userId": {
                    "type": "integer",
                    "description": "ID do usuário responsável"
                  }
                }
              },
              "examples": {
                "ticket": {
                  "value": {
                    "status": "pending",
                    "queueId": 2,
                    "userId": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket atualizado com sucesso"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'update:tickets'"
          },
          "404": {
            "description": "Ticket não encontrado"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      },
      "delete": {
        "summary": "Excluir Ticket",
        "description": "Excluir um ticket específico via API\n\nPermissões requeridas: `delete:tickets`",
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "apiToken": [
              "delete:tickets"
            ]
          }
        ],
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "ID do ticket",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket excluído com sucesso"
          },
          "401": {
            "description": "Token inválido ou não fornecido"
          },
          "403": {
            "description": "Token não tem permissão 'delete:tickets'"
          },
          "404": {
            "description": "Ticket não encontrado"
          },
          "500": {
            "description": "Erro interno!"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Ticket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do ticket"
          },
          "status": {
            "type": "string",
            "description": "Status do ticket (open, pending, closed)",
            "enum": ["open", "pending", "closed"]
          },
          "unreadMessages": {
            "type": "integer",
            "description": "Número de mensagens não lidas"
          },
          "lastMessage": {
            "type": "string",
            "description": "Conteúdo da última mensagem"
          },
          "contactId": {
            "type": "integer",
            "description": "ID do contato associado ao ticket"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact",
            "description": "Dados do contato associado ao ticket"
          },
          "queueId": {
            "type": "integer",
            "description": "ID do setor associado ao ticket"
          },
          "userId": {
            "type": "integer",
            "description": "ID do usuário responsável pelo ticket"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação do ticket"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data da última atualização do ticket"
          }
        }
      },
      "Contact": {
        "type": "object",
        "required": [
          "name",
          "number"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do contato"
          },
          "number": {
            "type": "string",
            "description": "Número do contato no formato DDI+DDD+NÚMERO (ex: 5522999999999)"
          },
          "email": {
            "type": "string",
            "description": "Email do contato"
          },
          "address": {
            "type": "string",
            "description": "Endereço do contato"
          },
          "extraInfo": {
            "type": "array",
            "description": "Informações adicionais do contato",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "Nome do campo adicional"
                },
                "value": {
                  "type": "string",
                  "description": "Valor do campo adicional"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "apiToken": {
        "type": "apiKey",
        "name": "x-api-token",
        "in": "header",
        "description": "Token de API para autenticação. Requer permissões específicas para cada rota:\n- `create:messages`: Para enviar mensagens\n- `create:contacts`: Para criar contatos"
      }
    }
  }
}